+++ c2dataviewer/__init__.py
@@ -1,2 +1,3 @@
 from .imagev import imagev
 from .scope import scope
+from .striptool import striptool