+++ c2dataviewer/view/__init__.py
@@ -1,2 +1,3 @@
 from .scopeconfig import Configure
+from .striptoolconfig import StripToolConfigure
 from .image_display import ImagePlotWidget